requirements.txt
scapy>=2.4.0
requests>=2.25.1
dnspython>=2.1.0
colorama>=0.4.4
tqdm>=4.60.0
